refactor: remove unused chart abstraction layer
- Removed unused abstraction layer components (ChartProvider, wrapper components)
- Removed dead recharts adapter code
- Simplified directory structure by moving Plotly components up one level
- Updated all import paths to reflect new structure
- Kept minimal types.ts for backward compatibility
- All builds tested and working (ui-components, GoPCA, GoCSV)

The abstraction layer was never used in production and added unnecessary
complexity. Direct Plotly usage is simpler and gives full access to features.
