+++ packages/ui-components/src/charts/PlotlyBarChart.tsx
@@ -6,9 +6,9 @@
 
 import React, { useEffect, useRef } from 'react';
 import Plotly from 'plotly.js-basic-dist-min';
-import { BarChartProps } from '../../types';
+import { BarChartProps } from './types';
 import { getPlotlyTheme, mergeLayouts } from './utils';
-import { useChartTheme } from '../../../hooks/useChartTheme';
+import { useChartTheme } from '../hooks/useChartTheme';
 
 export const PlotlyBarChart: React.FC<BarChartProps> = ({
   data,
