+++ packages/ui-components/src/charts/PlotlyScatterChart.tsx
@@ -6,9 +6,9 @@
 
 import React, { useEffect, useRef, useMemo } from 'react';
 import Plotly from 'plotly.js-basic-dist-min';
-import { ScatterChartProps } from '../../types';
+import { ScatterChartProps } from './types';
 import { getPlotlyTheme, mergeLayouts, calculatePlotlyLabels, getPlotlyTextPosition } from './utils';
-import { useChartTheme } from '../../../hooks/useChartTheme';
+import { useChartTheme } from '../hooks/useChartTheme';
 
 export const PlotlyScatterChart: React.FC<ScatterChartProps> = ({
   data,
