+++ packages/ui-components/src/charts/core/PlotlyVisualization.tsx
