+++ packages/ui-components/src/charts/pca/Plotly3DScoresPlot.tsx
