+++ packages/ui-components/src/charts/pca/PlotlyDiagnosticPlot.tsx
