+++ packages/ui-components/src/charts/pca/PlotlyScoresPlot.tsx
