+++ packages/ui-components/src/charts/pca/PlotlyScreePlot.tsx
