+++ packages/ui-components/src/charts/utils/plotlyHelpers.ts
