+++ packages/ui-components/src/charts/utils/plotlyPerformance.ts
