+++ packages/ui-components/src/charts/utils/plotlyTheme.ts
